This commit is to create the basics of HTTP request parsing and function handling.
Added HttpRequest class that takes in a vector<char> and parses the HttpRequest into the fields necessary

Changed our config parsing to map directories to function names. If we see a get request with a directory, we now know what function they want ("/echo" means that they want and echo back for example).

We still need to make our code more modular and separate a class for sending an echo from one that sends a static file. We also need an HTTP response construction class.